fixed user deletion issue

+++ service/src/main/java/com/soen343/project/service/catalog/Catalog.java
@@ -22,6 +22,7 @@ import static com.soen343.project.database.connection.DatabaseConnector.executeQ
 import static com.soen343.project.repository.dao.catalog.item.com.LoanableItemOperation.findAllLoanables;
 import static com.soen343.project.repository.dao.catalog.item.com.LoanableItemOperation.queryLoanableAndItemByItemspecType;
 import static com.soen343.project.repository.entity.EntityConstants.ID;
+import static com.soen343.project.repository.entity.user.types.UserType.ADMIN;
 
 @Service
 public class Catalog {
@@ -96,11 +97,15 @@ public class Catalog {
 
     public ResponseEntity<?> deleteUser(String sessionID, Long userId) {
         CatalogSession session = getSession(sessionID);
-        if (library.getAllAdmins().size() > 1) {
-            session.deleteUser(userId);
-            return new ResponseEntity<>(HttpStatus.OK);
+        if (library.getUserByID(userId).getUserType().equals(ADMIN)) {
+            if (library.getAllAdmins().size() > 1) {
+                session.deleteUser(userId);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.PRECONDITION_FAILED);
+            }
         } else {
-            return new ResponseEntity<>(HttpStatus.PRECONDITION_FAILED);
+            return new ResponseEntity<>(HttpStatus.OK);
         }
     }
